Fixed the annoying line on the axes and changed it to compute a colour for every single pixel.

+++ include/advanced_cxx/img_scaler.h
@@ -5,20 +5,20 @@
 
 namespace mabz {
 
-// Used to map from one scale to another, e.g. if you have an x,y range of 
-// a cartesian plane / graph kind of thing and you want to turn it into an 800 x 800 image...
+// Used to map from one scale to another, e.g. if you have x,y pixel coordinates in
+// an 800 x 600 image and you want to map them to an (x,y) cartesian plane from [-1, +1]
 // Could later be turned into a generic to map from type A to type B.
 class ImgScaler
 {
 private:
-	const double mSourceMin{0};
-	const double mSourceRange{0};
-	const int mDestMin{0};
-	const int mDestRange{0};
+	const int mSourceMin{0};
+	const int mSourceRange{0};
+	const double mDestMin{0};
+	const double mDestRange{0};
 
 public:
 	ImgScaler() = delete;
-	ImgScaler(double sourceMin, double sourceMax, int destMin, int destMax)
+	ImgScaler(int sourceMin, int sourceMax, double destMin, double destMax)
 		: mSourceMin(sourceMin)
 		, mSourceRange(sourceMax - sourceMin)
 		, mDestMin(destMin)
@@ -32,7 +32,7 @@ public:
 
 	// Simply a linear extrapolation based on distance from the midpoint of 
 	// the range; does not bother to check whether the provided 'in' value is within source range!!
-	void Convert(double in, int& out) const;
+	void Convert(int in, double& out) const;
 };
 
 } /* namespace mabz */
